third_party/WebKit/LayoutTests/fast/canvas-api/OffscreenCanvas-invalid-args-in-worker-expected.txt

Tests that the OffscreenCanvas can handle invalid arguments on a worker

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS width is setWidth-1
PASS height is setHeight-1
PASS width is 0
PASS height is 0
PASS successfullyParsed is true

TEST COMPLETE
